Make Metax integration optional.

+++ metadata_backend/api/handlers/object.py
@@ -185,13 +185,21 @@ class ObjectAPIHandler(RESTAPIHandler):
         patch = self._prepare_submission_patch_new_object(collection, objects, cont_type)
         await submission_op.update_submission(submission_id, patch)
 
+        # Add DOI to object
+        if collection in _allowed_doi:
+            obj: Dict[str, Any]
+            for obj, _ in objects:
+                obj["doi"] = await self.create_draft_doi(collection)
+                await Operator(db_client).create_datacite_info(collection, obj["accessionId"], {"doi": obj["doi"]})
+
         # Create draft dataset to Metax catalog
+        metax_handler = MetaxServiceHandler(req)
         try:
-            if collection in _allowed_doi:
+            if metax_handler.enabled and collection in _allowed_doi:
                 [await self.create_metax_dataset(req, collection, item) for item, _ in objects]
         except Exception as e:
             # We don't care if it fails here
-            LOG.info(f"create_metax_dataset failed: {e} for collection {collection}")
+            LOG.exception(f"create_metax_dataset failed: {e} for collection {collection}")
             pass
 
         body = ujson.dumps(data, escape_forward_slashes=False)
@@ -266,12 +274,13 @@ class ObjectAPIHandler(RESTAPIHandler):
         accession_id = await operator.delete_metadata_object(collection, accession_id)
 
         # Delete draft dataset from Metax catalog
-        if collection in _allowed_doi:
+        metax_handler = MetaxServiceHandler(req)
+        if metax_handler.enabled and collection in _allowed_doi:
             try:
                 await MetaxServiceHandler(req).delete_draft_dataset(metax_id)
             except Exception as e:
                 # We don't care if it fails here
-                LOG.info(f"delete_draft_dataset failed: {e} for collection {collection}")
+                LOG.exception(f"delete_draft_dataset failed: {e} for collection {collection}")
                 pass
             doi_service = DataciteServiceHandler()
             await doi_service.delete(doi_id)
@@ -330,15 +339,16 @@ class ObjectAPIHandler(RESTAPIHandler):
         await submission_op.update_submission(submission_id, patch)
 
         # Update draft dataset to Metax catalog
+        metax_handler = MetaxServiceHandler(req)
         try:
-            if collection in _allowed_doi:
+            if metax_handler.enabled and collection in _allowed_doi:
                 if data.get("metaxIdentifier", None):
                     await MetaxServiceHandler(req).update_draft_dataset(collection, data)
                 else:
-                    await self.create_metax_dataset(req, collection, data, create_draft_doi=False)
+                    await self.create_metax_dataset(req, collection, data)
         except Exception as e:
             # We don't care if it fails here
-            LOG.info(f"update_draft_dataset or create_metax_dataset failed: {e} for collection {collection}")
+            LOG.exception(f"update_draft_dataset or create_metax_dataset failed: {e} for collection {collection}")
             pass
 
         body = ujson.dumps({"accessionId": accession_id}, escape_forward_slashes=False)
@@ -393,20 +403,21 @@ class ObjectAPIHandler(RESTAPIHandler):
             pass
 
         # Update draft dataset to Metax catalog
+        metax_handler = MetaxServiceHandler(req)
         try:
-            if collection in {"study", "dataset"}:
+            if metax_handler.enabled and collection in {"study", "dataset"}:
                 object_data, _ = await operator.read_metadata_object(collection, accession_id)
                 # MYPY related if statement, Operator (when not XMLOperator) always returns object_data as dict
                 if isinstance(object_data, Dict):
                     if object_data.get("metaxIdentifier", None):
                         await MetaxServiceHandler(req).update_draft_dataset(collection, object_data)
                     else:
-                        await self.create_metax_dataset(req, collection, object_data, create_draft_doi=False)
+                        await self.create_metax_dataset(req, collection, object_data)
                 else:
                     raise ValueError("Object's data must be dictionary")
         except Exception as e:
             # We don't care if it fails here
-            LOG.info(f"update_draft_dataset or create_metax_dataset failed: {e} for collection {collection}")
+            LOG.exception(f"update_draft_dataset or create_metax_dataset failed: {e} for collection {collection}")
             pass
 
         body = ujson.dumps({"accessionId": accession_id}, escape_forward_slashes=False)
@@ -504,9 +515,8 @@ class ObjectAPIHandler(RESTAPIHandler):
 
         return [patch_op, lastModified]
 
-    async def create_metax_dataset(
-        self, req: Request, collection: str, object: Dict, create_draft_doi: bool = True
-    ) -> str:
+    @staticmethod
+    async def create_metax_dataset(req: Request, collection: str, object: Dict) -> str:
         """Handle connection to Metax api handler for dataset creation.
 
         Dataset or Study object is assigned with DOI
@@ -516,32 +526,35 @@ class ObjectAPIHandler(RESTAPIHandler):
         :param req: HTTP request
         :param collection: object's schema
         :param object: metadata object
-        :param submission_id: submission ID where metadata object belongs to
         :returns: Metax ID
         """
         LOG.info("Creating draft dataset to Metax.")
+        metax_handler = MetaxServiceHandler(req)
+        if not metax_handler.enabled:
+            LOG.error("Metax integration is disabled.")
+            return ""
         operator = Operator(req.app["db_client"])
         new_info = {}
-        if create_draft_doi:
-            object["doi"] = await self._draft_doi(collection)
-            new_info = {"doi": object["doi"]}
-        metax_id = await MetaxServiceHandler(req).post_dataset_as_draft(collection, object)
+        if "doi" in object:
+            new_info["doi"] = object["doi"]
+        metax_id = await metax_handler.post_dataset_as_draft(collection, object)
         new_info["metaxIdentifier"] = metax_id
-        await operator.create_metax_info(collection, object["accessionId"], new_info)
+        await operator.create_datacite_info(collection, object["accessionId"], new_info)
 
         return metax_id
 
-    async def _draft_doi(self, schema_type: str) -> str:
+    @staticmethod
+    async def create_draft_doi(collection: str) -> dict:
         """Create draft DOI for study and dataset.
 
         The Draft DOI will be created only on POST and the data added to the
         submission. Any update of this should not be possible.
 
-        :param schema_type: schema can be either study or dataset
+        :param collection: schema can be either study or dataset
         :returns: Dict with DOI of the study or dataset as well as the types.
         """
         doi_ops = DataciteServiceHandler()
-        _doi_data = await doi_ops.create_draft(prefix=schema_type)
+        _doi_data = await doi_ops.create_draft(prefix=collection)
 
         LOG.debug(f"doi created with doi: {_doi_data['fullDOI']}")
